refactor: reorganize SpotifyAuth package structure and remove unused classes
