+++ src/main/java/app/AppBuilder.java
@@ -10,9 +10,9 @@ import data_access.InMemoryPlaylistDataAccessObject;
 import data_access.InMemoryUserDataAccessObject;
 import entity.Track;
 import entity.User;
-import interface_adapter.SpotifyAuth.LoginController;
-import interface_adapter.SpotifyAuth.LoginPresenter;
-import interface_adapter.SpotifyAuth.LoginViewModel;
+import interface_adapter.spotify_auth.LoginController;
+import interface_adapter.spotify_auth.LoginPresenter;
+import interface_adapter.spotify_auth.LoginViewModel;
 import interface_adapter.loggedin.LoggedInViewModel;
 import interface_adapter.loggedin.LoggedInView;
 import interface_adapter.ViewManagerModel;
@@ -30,6 +30,9 @@ import interface_adapter.rec_playlist.RecPlaylistViewModel;
 import interface_adapter.rec_song.RecSongController;
 import interface_adapter.rec_song.RecSongPresenter;
 import interface_adapter.rec_song.RecSongViewModel;
+import interface_adapter.spotify_auth.LoginController;
+import interface_adapter.spotify_auth.LoginPresenter;
+import interface_adapter.spotify_auth.LoginViewModel;
 import use_case.login.LoginInputBoundary;
 import use_case.login.LoginInteractor;
 import use_case.login.LoginOutputBoundary;
