+++ src/main/java/interface_adapter/spotify_auth/LoginViewModel.java
@@ -1,4 +1,6 @@
-package interface_adapter.spotifyauth
+package interface_adapter.spotify_auth;
+
+
 
 import interface_adapter.ViewModel;
