+++ src/main/java/interface_adapter/spotify_auth/SpotifyConfig.java
@@ -1,4 +1,4 @@
-package interface_adapter.spotifyauth;
+package interface_adapter.spotify_auth;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
