--- src/main/java/interface_adapter/spotifyauth/LoginController.java
@@ -1,28 +1,0 @@
-package interface_adapter.SpotifyAuth
-
-import use_case.login.LoginInputBoundary;
-import use_case.login.LoginInputData;
-
-/**
- * The controller for the Login Use Case.
- */
-public class LoginController {
-
-    private final LoginInputBoundary loginUseCaseInteractor;
-
-    public LoginController(LoginInputBoundary loginUseCaseInteractor) {
-        this.loginUseCaseInteractor = loginUseCaseInteractor;
-    }
-
-    /**
-     * Executes the Login Use Case.
-     * @param username the username of the user logging in
-     * @param password the password of the user logging in
-     */
-    public void execute(String username, String password) {
-        final LoginInputData loginInputData = new LoginInputData(
-                username, password);
-
-        loginUseCaseInteractor.execute(loginInputData);
-    }
-}
