--- src/main/java/interface_adapter/spotifyauth/LoginViewModel.java
@@ -1,17 +1,0 @@
-package interface_adapter.SpotifyAuth;
-
-
-import interface_adapter.SpotifyAuth.LoginState;
-import interface_adapter.ViewModel;
-
-/**
- * The View Model for the Login View.
- */
-public class LoginViewModel extends ViewModel<LoginState> {
-
-    public LoginViewModel() {
-        super("log in");
-        setState(new LoginState());
-    }
-
-}
