--- src/main/java/interface_adapter/spotifyauth/SpotifyConfig.java
@@ -1,29 +1,0 @@
-package interface_adapter.SpotifyAuth;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SpotifyConfig {
-
-    @Value("${spotify.client-id}")
-    private String clientId;
-
-    @Value("${spotify.redirect-uri}")
-    private String redirectUri;
-
-    @Value("${spotify.scope}")
-    private String scope;
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public String getRedirectUri() {
-        return redirectUri;
-    }
-
-    public String getScope() {
-        return scope;
-    }
-}
